+++ pkg/koyeb/organizations_list.go
@@ -11,9 +11,19 @@ import (
 )
 
 func getCurrentUserId(ctx *CLIContext) (string, error) {
-	res, resp, err := ctx.Client.ProfileApi.GetCurrentUser(ctx.Context).Execute()
+	res, _, err := ctx.Client.ProfileApi.GetCurrentUser(ctx.Context).Execute()
 	if err != nil {
-		return "", errors.NewCLIErrorFromAPIError("The token used is not linked to a user", err, resp)
+		return "", &errors.CLIError{
+			What: "The token used is not linked to a user",
+			Why:  "you are authenticated with a token linked to an organization",
+			Additional: []string{
+				"On Koyeb, two types of tokens exist: user tokens and organization tokens.",
+				"Your are currently using an organization token, which is not linked to a user.",
+				"Organization tokens are unable to perform operations that require a user context, such as listing organizations or managing your account.",
+			},
+			Orig:     err,
+			Solution: "From the Koyeb console (https://app.koyeb.com/user/settings/api/), create a user token and use it in the CLI configuration file.",
+		}
 	}
 	return *res.GetUser().Id, nil
 }